# apps/piteka-api/src/handlers.rs
//! Request handlers for the first-slice API endpoints.

use axum::{
    Json,
    extract::{FromRequestParts, Path, State},
    http::{StatusCode, request::Parts},
    response::{IntoResponse, Response},
};
use piteka_application::ActionRequestUseCase;
use piteka_application::bundle_export::assemble_bundle;
use piteka_domain::UserId;
use piteka_storage::ports::ReceiptProjectionStore;

use crate::TestPorts;
use crate::error::{ApiError, ErrorResponseV1};
use crate::models::{
    ActionRequestResponseV1, ActionRequestSummaryDtoV1, ApproveActionRequestRequestV1,
    CreateActionRequestRequestV1, RejectActionRequestRequestV1, RevokeActionRequestRequestV1,
};

/// Optional idempotency key extracted from the request header.
#[derive(Debug, Default)]
pub struct IdempotencyKey(pub Option<String>);

impl<S> FromRequestParts<S> for IdempotencyKey
where
    S: Send + Sync,
{
    type Rejection = (StatusCode, Json<ErrorResponseV1>);

    async fn from_request_parts(parts: &mut Parts, _state: &S) -> Result<Self, Self::Rejection> {
        let key = parts
            .headers
            .get("Idempotency-Key")
            .and_then(|v| v.to_str().ok())
            .map(|s| s.to_string());
        Ok(IdempotencyKey(key))
    }
}

/// List all action requests.
///
/// `GET /api/v1/action-requests`
///
/// Returns a 200 with an array of summaries.
pub async fn list_action_requests(
    State(use_case): State<ActionRequestUseCase<TestPorts>>,
) -> Response {
    match use_case.list_requests().await {
        Ok(requests) => {
            let summaries: Vec<ActionRequestSummaryDtoV1> = requests
                .into_iter()
                .map(|r| ActionRequestSummaryDtoV1 {
                    id: r.request_id,
                    requested_by: r.requested_by,
                    status: r.status.into(),
                    created_at: r.created_at_unix_seconds as u64,
                })
                .collect();
            Json(summaries).into_response()
        }
        Err(err) => ApiError::from(err).into_response(),
    }
}

/// Get a single action request by id.
///
/// `GET /api/v1/action-requests/{id}`
///
/// Returns the full request detail including decisions, or 404 if not found.
pub async fn get_action_request(
    State(use_case): State<ActionRequestUseCase<TestPorts>>,
    Path(request_id): Path<String>,
) -> Response {
    let request = match use_case.get_request(&request_id).await {
        Ok(r) => r,
        Err(err) => return ApiError::from(err).into_response(),
    };

    let request = match request {
        Some(r) => r,
        None => return ApiError::not_found("action request", &request_id).into_response(),
    };

    let decisions = match use_case.get_decisions(&request_id).await {
        Ok(d) => d,
        Err(err) => return ApiError::from(err).into_response(),
    };

    let response = ActionRequestResponseV1 {
        id: request.request_id,
        requested_by: request.requested_by,
        intent_id: request.intent_id_hex,
        status: request.status.into(),
        created_at: request.created_at_unix_seconds as u64,
        decisions: decisions.into_iter().map(|d| d.into()).collect(),
    };

    Json(response).into_response()
}

/// Propose a new action request.
///
/// `POST /api/v1/action-requests`
///
/// Supports idempotency via the `Idempotency-Key` header. A second call with
/// the same key returns the original result.
pub async fn propose_action_request(
    State(use_case): State<ActionRequestUseCase<TestPorts>>,
    _idempotency_key: IdempotencyKey,
    Json(body): Json<CreateActionRequestRequestV1>,
) -> Response {
    // Validate request body
    if body.requested_by.is_empty() {
        return ApiError::bad_request("EMPTY_REQUESTED_BY", "The `requested_by` field is required")
            .into_response();
    }

    let user_id = match UserId::new(&body.requested_by) {
        Ok(id) => id,
        Err(_) => UserId::new("unknown").unwrap(),
    };

    let request_id = uuid::Uuid::new_v4().to_string();

    let proposed = match use_case.propose(&request_id, user_id, body.intent_id).await {
        Ok(p) => p,
        Err(err) => return ApiError::from(err).into_response(),
    };

    let request = proposed.request;
    let response = ActionRequestResponseV1 {
        id: request.request_id,
        requested_by: request.requested_by,
        intent_id: request.intent_id_hex,
        status: request.status.into(),
        created_at: request.created_at_unix_seconds as u64,
        decisions: Vec::new(),
    };

    (StatusCode::CREATED, Json(response)).into_response()
}

/// Approve an action request.
///
/// `POST /api/v1/action-requests/{id}/approve`
///
/// Supports idempotency via the `Idempotency-Key` header.
pub async fn approve_action_request(
    State(use_case): State<ActionRequestUseCase<TestPorts>>,
    _idempotency_key: IdempotencyKey,
    Path(request_id): Path<String>,
    Json(body): Json<ApproveActionRequestRequestV1>,
) -> Response {
    if body.approver_id.is_empty() {
        return ApiError::bad_request("EMPTY_APPROVER_ID", "The `approver_id` field is required")
            .into_response();
    }

    let user_id = match UserId::new(&body.approver_id) {
        Ok(id) => id,
        Err(_) => UserId::new("unknown").unwrap(),
    };

    let approved = match use_case
        .approve(&request_id, user_id, body.intent_id, body.version)
        .await
    {
        Ok(a) => a,
        Err(err) => return ApiError::from(err).into_response(),
    };

    let decisions = match use_case.get_decisions(&request_id).await {
        Ok(d) => d,
        Err(err) => return ApiError::from(err).into_response(),
    };

    let response = ActionRequestResponseV1 {
        id: approved.request.request_id,
        requested_by: approved.request.requested_by,
        intent_id: approved.request.intent_id_hex,
        status: approved.request.status.into(),
        created_at: approved.request.created_at_unix_seconds as u64,
        decisions: decisions.into_iter().map(|d| d.into()).collect(),
    };

    Json(response).into_response()
}

/// Reject an action request.
///
/// `POST /api/v1/action-requests/{id}/reject`
///
/// Supports idempotency via the `Idempotency-Key` header.
pub async fn reject_action_request(
    State(use_case): State<ActionRequestUseCase<TestPorts>>,
    _idempotency_key: IdempotencyKey,
    Path(request_id): Path<String>,
    Json(body): Json<RejectActionRequestRequestV1>,
) -> Response {
    if body.approver_id.is_empty() {
        return ApiError::bad_request("EMPTY_APPROVER_ID", "The `approver_id` field is required")
            .into_response();
    }

    let user_id = match UserId::new(&body.approver_id) {
        Ok(id) => id,
        Err(_) => UserId::new("unknown").unwrap(),
    };

    let rejected = match use_case
        .reject(&request_id, user_id, body.intent_id, body.version)
        .await
    {
        Ok(r) => r,
        Err(err) => return ApiError::from(err).into_response(),
    };

    let decisions = match use_case.get_decisions(&request_id).await {
        Ok(d) => d,
        Err(err) => return ApiError::from(err).into_response(),
    };

    let response = ActionRequestResponseV1 {
        id: rejected.request.request_id,
        requested_by: rejected.request.requested_by,
        intent_id: rejected.request.intent_id_hex,
        status: rejected.request.status.into(),
        created_at: rejected.request.created_at_unix_seconds as u64,
        decisions: decisions.into_iter().map(|d| d.into()).collect(),
    };

    Json(response).into_response()
}

/// Revoke an approved action request.
///
/// `POST /api/v1/action-requests/{id}/revoke`
///
/// Supports idempotency via the `Idempotency-Key` header.
pub async fn revoke_action_request(
    State(use_case): State<ActionRequestUseCase<TestPorts>>,
    _idempotency_key: IdempotencyKey,
    Path(request_id): Path<String>,
    Json(body): Json<RevokeActionRequestRequestV1>,
) -> Response {
    if body.approver_id.is_empty() {
        return ApiError::bad_request("EMPTY_APPROVER_ID", "The `approver_id` field is required")
            .into_response();
    }

    let user_id = match UserId::new(&body.approver_id) {
        Ok(id) => id,
        Err(_) => UserId::new("unknown").unwrap(),
    };

    let revoked = match use_case.revoke(&request_id, user_id, body.version).await {
        Ok(r) => r,
        Err(err) => return ApiError::from(err).into_response(),
    };

    let decisions = match use_case.get_decisions(&request_id).await {
        Ok(d) => d,
        Err(err) => return ApiError::from(err).into_response(),
    };

    let response = ActionRequestResponseV1 {
        id: revoked.request.request_id,
        requested_by: revoked.request.requested_by,
        intent_id: revoked.request.intent_id_hex,
        status: revoked.request.status.into(),
        created_at: revoked.request.created_at_unix_seconds as u64,
        decisions: decisions.into_iter().map(|d| d.into()).collect(),
    };

    Json(response).into_response()
}

/// Get a receipt by id.
///
/// `GET /api/v1/receipts/{id}`
///
/// Returns the receipt projection including source attribution and evidence
/// references, or 404 if not found.
pub async fn get_receipt(
    State(ports): State<TestPorts>,
    Path(receipt_id): Path<String>,
) -> Response {
    match ports.receipt_store.get(&ports.tenant, &receipt_id).await {
        Ok(Some(receipt)) => {
            let response = serde_json::json!({
                "receipt_id": receipt.receipt_id_hex,
                "mandate_id": receipt.mandate_id_hex,
                "intent_id": receipt.intent_id_hex,
                "attempt_id": receipt.attempt_id_hex,
                "outcome": match receipt.outcome {
                    piteka_storage::model::ReceiptOutcome::Succeeded => "succeeded",
                    piteka_storage::model::ReceiptOutcome::Failed => "failed",
                    piteka_storage::model::ReceiptOutcome::Rejected => "rejected",
                    piteka_storage::model::ReceiptOutcome::Unknown => "unknown",
                },
                "created_at": receipt.created_at_unix_seconds,
                "dispatch_evidence_refs": receipt.dispatch_evidence_refs,
                "target_evidence_refs": receipt.target_evidence_refs,
                "evidence_gaps": receipt.evidence_gaps,
            });
            Json(response).into_response()
        }
        Ok(None) => ApiError::not_found("receipt", &receipt_id).into_response(),
        Err(err) => ApiError::internal(format!("storage error: {err}")).into_response(),
    }
}

/// Export a DisputeBundle for a receipt.
///
/// `GET /api/v1/receipts/{id}/export`
///
/// Assembles a portable case file containing the receipt, evidence nodes,
/// and gaps. Returns the bundle manifest as JSON.
pub async fn export_bundle(
    State(ports): State<TestPorts>,
    Path(receipt_id): Path<String>,
) -> Response {
    match assemble_bundle(
        &ports.tenant,
        &ports.receipt_store,
        &ports.evidence_store,
        &ports.evidence_blob_store,
        &ports.protocol_store,
        &ports.seal_consumption_store,
        &receipt_id,
    )
    .await
    {
        Ok(bundle) => {
            let response = serde_json::json!({
                "bundle_id": bundle.bundle_id_hex,
                "receipt_id": bundle.receipt_id_hex,
                "mandate_id": bundle.mandate_id_hex,
                "evidence_node_ids": bundle.evidence_node_ids,
                "evidence_gap_ids": bundle.evidence_gap_ids,
                "bundle_digest": bundle.bundle_digest.to_hex(),
            });
            Json(response).into_response()
        }
        Err(err) => ApiError::internal(format!("bundle export failed: {err}")).into_response(),
    }
}

/// Get the canonical receipt bytes for a receipt.
///
/// `GET /api/v1/receipts/{id}/canonical`
///
/// Returns the Parwana-canonical bytes of the receipt, if already serialized.
pub async fn get_receipt_canonical(
    State(ports): State<TestPorts>,
    Path(receipt_id): Path<String>,
) -> Response {
    match ports.receipt_store.get(&ports.tenant, &receipt_id).await {
        Ok(Some(receipt)) => match receipt.canonical_bytes {
            Some(bytes) => (StatusCode::OK, bytes).into_response(),
            None => ApiError::not_found("canonical receipt bytes", &receipt_id).into_response(),
        },
        Ok(None) => ApiError::not_found("receipt", &receipt_id).into_response(),
        Err(err) => ApiError::internal(format!("storage error: {err}")).into_response(),
    }
}